requirements.txt
pypdf
sentence-transformers
faiss-cpu
transformers
torch
numpy
tqdm
